Move track page selectors into common
