+++ src/pages/track-page/TrackPageProvider.tsx
@@ -19,6 +19,15 @@ import { Track } from 'common/models/Track'
 import { getUserId } from 'common/store/account/selectors'
 import * as cacheTrackActions from 'common/store/cache/tracks/actions'
 import { makeGetLineupMetadatas } from 'common/store/lineup/selectors'
+import {
+  getUser,
+  getLineup,
+  getTrackRank,
+  getTrack,
+  getRemixParentTrack,
+  getStatus,
+  getSourceSelector
+} from 'common/store/pages/track/selectors'
 import * as socialTracksActions from 'common/store/social/tracks/actions'
 import * as socialUsersActions from 'common/store/social/users/actions'
 import { open } from 'common/store/ui/mobile-overflow-menu/slice'
@@ -34,15 +43,6 @@ import DeletedPage from 'pages/deleted-page/DeletedPage'
 import { setFavorite } from 'pages/favorites-page/store/actions'
 import { setRepost } from 'pages/reposts-page/store/actions'
 import { RepostType } from 'pages/reposts-page/store/types'
-import {
-  getUser,
-  getLineup,
-  getTrackRank,
-  getTrack,
-  getRemixParentTrack,
-  getStatus,
-  getSourceSelector
-} from 'pages/track-page/store/selectors'
 import { TrackEvent, make } from 'store/analytics/actions'
 import {
   setUsers,
