+++ src/pages/track-page/store/lineups/tracks/sagas.js
@@ -3,11 +3,11 @@ import { call, select } from 'redux-saga/effects'
 import { getUserId } from 'common/store/account/selectors'
 import { getTrack } from 'common/store/cache/tracks/selectors'
 import { PREFIX, tracksActions } from 'common/store/pages/track/lineup/actions'
-import { retrieveUserTracks } from 'pages/profile-page/store/lineups/tracks/retrieveUserTracks'
 import {
   getSourceSelector as sourceSelector,
   getLineup
-} from 'pages/track-page/store/selectors'
+} from 'common/store/pages/track/selectors'
+import { retrieveUserTracks } from 'pages/profile-page/store/lineups/tracks/retrieveUserTracks'
 import { LineupSagas } from 'store/lineup/sagas'
 import { waitForValue } from 'utils/sagaHelpers'
