+++ src/store/lineup/lineupForRoute.js
@@ -1,12 +1,12 @@
 import { matchPath } from 'react-router'
 
+import { getLineup } from 'common/store/pages/track/selectors'
 import { getCollectionTracksLineup } from 'pages/collection-page/store/selectors'
 import { getDiscoverFeedLineup } from 'pages/feed-page/store/selectors'
 import { getHistoryTracksLineup } from 'pages/history-page/store/selectors'
 import { getProfileTracksLineup } from 'pages/profile-page/store/selectors'
 import { getSavedTracksLineup } from 'pages/saved-page/store/selectors'
 import { getSearchTracksLineup } from 'pages/search-page/store/selectors'
-import { getLineup } from 'pages/track-page/store/selectors'
 import { getCurrentDiscoverTrendingLineup } from 'pages/trending-page/store/selectors'
 import {
   FEED_PAGE,
